修正倍率为提升数值 而不是成倍增长

+++ 游戏库/libpsocnconfig/psoconfig.h
@@ -262,9 +262,9 @@ typedef struct psocn_shipcfg {
     char* gm_file;
     psocn_limit_config_t* limits;
     uint32_t globla_exp_mult;
-    uint32_t monster_rare_mult;
-    uint32_t monster_rare_drop_mult;
-    uint32_t box_rare_drop_mult;
+    double monster_rare_mult;
+    double monster_rare_drop_mult;
+    double box_rare_drop_mult;
     psocn_info_file_t* info_files;
     char* quests_file;
     char* quests_dir;
